Rename Vert.x Template to PRL Harvester
