+++ src/main/java/edu/ucla/library/prl/harvester/handlers/StatusHandler.java
@@ -1,11 +1,11 @@
 
-package info.freelibrary.vertx.template.handlers;
+package edu.ucla.library.prl.harvester.handlers;
 
-import static info.freelibrary.vertx.template.MediaType.APPLICATION_JSON;
+import static edu.ucla.library.prl.harvester.MediaType.APPLICATION_JSON;
+
+import edu.ucla.library.prl.harvester.JsonKeys;
 
 import info.freelibrary.util.HTTP;
-
-import info.freelibrary.vertx.template.JsonKeys;
 
 import io.vertx.core.Handler;
 import io.vertx.core.Vertx;
