+++ src/main/resources/META-INF/native-image/info.freelibrary/prl-harvester/jni-config.json
