+++ src/main/resources/META-INF/native-image/info.freelibrary/prl-harvester/predefined-classes-config.json
