+++ src/main/resources/META-INF/native-image/info.freelibrary/prl-harvester/proxy-config.json
