+++ src/main/resources/META-INF/native-image/info.freelibrary/prl-harvester/reflect-config.json
@@ -715,7 +715,7 @@
     "allDeclaredMethods": true
   },
   {
-    "name": "info.freelibrary.vertx.template.verticles.MainVerticle",
+    "name": "edu.ucla.library.prl.harvester.verticles.MainVerticle",
     "allDeclaredConstructors": true,
     "allDeclaredMethods": true
   }
