+++ src/main/resources/META-INF/native-image/info.freelibrary/prl-harvester/resource-config.json
