+++ src/main/resources/META-INF/native-image/info.freelibrary/prl-harvester/serialization-config.json
