+++ src/main/resources/prl-harvester_messages.xml
@@ -1,0 +1,12 @@
+<?xml version="1.0" encoding="UTF-8"?>
+
+<!DOCTYPE properties SYSTEM "http://java.sun.com/dtd/properties.dtd">
+
+<properties>
+  <entry key="message-class-name">edu.ucla.library.prl.harvester.MessageCodes</entry>
+
+  <entry key="PRL_001">Server started at port: {}</entry>
+  <entry key="PRL_002"></entry>
+  <entry key="PRL_003"></entry>
+
+</properties>
