--- src/main/resources/template_messages.xml
@@ -1,12 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-
-<!DOCTYPE properties SYSTEM "http://java.sun.com/dtd/properties.dtd">
-
-<properties>
-  <entry key="message-class-name">info.freelibrary.vertx.template.MessageCodes</entry>
-
-  <entry key="CODE_001">Server started at port: {}</entry>
-  <entry key="CODE_002"></entry>
-  <entry key="CODE_003"></entry>
-
-</properties>
