+++ src/test/resources/logback-test.xml
@@ -9,7 +9,7 @@
   </appender>
 
   <appender name="FILE" class="ch.qos.logback.core.FileAppender">
-    <file>target/vertx-template-${byDay}.log</file>
+    <file>target/prl-harvester-${byDay}.log</file>
     <append>false</append>
     <encoder>
       <pattern>[%level] %logger{45} | %msg%n</pattern>
